makes lights work in LightSystem and makes Teleop thing to test LightSystem is kind of ugly but you know
